refactor(runtime): unify RouterDispatchHandler and NestedRoutingHandler into RoutingHandler
RouterDispatchHandler conflated four responsibilities: Netty handler
mechanics, router invocation, frame dispatch, and response correlation.
When nesting was added, NestedRoutingHandler reimplemented most of
this, and NestedRouterDispatch reimplemented the dispatch logic with
the only real difference being route name qualification.

Extract dispatch and response correlation into RouteDispatcher,
parameterised by a route prefix. Replace both handler classes with a
single RoutingHandler that wraps one Router and is configured for
policy differences (response sequencing, close semantics, OOB handling,
CCSM interaction) via topLevel() and nested() factory methods.

Also fixes a latent bug where NestedRouterDispatch.sendToSpecificNode()
always set targetVirtualNodeId, even for routes targeting a nested
router. RouteDispatcher now consistently uses the targetsCluster()
guard.
