+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/routing/RouteDispatcher.java
@@ -1,0 +1,342 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.kroxylicious.proxy.internal.routing;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+import java.util.function.Supplier;
+
+import org.apache.kafka.common.message.MetadataResponseData;
+import org.apache.kafka.common.message.RequestHeaderData;
+import org.apache.kafka.common.protocol.ApiKeys;
+import org.apache.kafka.common.protocol.ApiMessage;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.spi.LoggingEventBuilder;
+
+import io.netty.channel.ChannelHandlerContext;
+
+import io.kroxylicious.proxy.frame.DecodedRequestFrame;
+import io.kroxylicious.proxy.frame.DecodedResponseFrame;
+import io.kroxylicious.proxy.internal.CorrelationIdAllocator;
+import io.kroxylicious.proxy.service.HostPort;
+
+import edu.umd.cs.findbugs.annotations.Nullable;
+
+/**
+ * Creates request frames and fires them downstream, tracks them by
+ * correlation ID, and matches returning response frames back to their
+ * pending futures — translating node IDs and caching upstream broker
+ * addresses along the way.
+ *
+ * <p>Parameterised by a route prefix (empty for the top-level router,
+ * {@code "routerName/"} for a nested router) so the same logic serves
+ * every router in the routing tree.
+ *
+ * <p>Not thread-safe; all callers must be on the same Netty event loop.
+ * Off-loop calls are bridged via {@link #executeOnEventLoop}.
+ */
+public class RouteDispatcher implements RouterDispatch {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(RouteDispatcher.class);
+
+    /**
+     * API keys whose responses carry node IDs that must be translated to virtual node IDs.
+     * These keys are always decoded (even when statically routed) so the response bodies
+     * are accessible for translation.
+     */
+    public static final Set<ApiKeys> NODE_ID_TRANSLATION_APIS = Set.of(
+            ApiKeys.METADATA,
+            ApiKeys.FIND_COORDINATOR,
+            ApiKeys.DESCRIBE_CLUSTER,
+            ApiKeys.PRODUCE,
+            ApiKeys.FETCH,
+            ApiKeys.SHARE_FETCH,
+            ApiKeys.SHARE_ACKNOWLEDGE,
+            ApiKeys.DESCRIBE_TOPIC_PARTITIONS);
+
+    private final Map<String, RouteDescriptor> routes;
+    private final NodeIdMapping nodeIdMapping;
+    private final String routePrefix;
+    private final CorrelationIdAllocator correlationIdAllocator;
+    private final Map<Integer, HostPort> routerNodeAddresses;
+    private final String virtualClusterName;
+
+    final Map<Integer, PendingResponse> pendingResponses = new HashMap<>();
+    final Map<Integer, String> pendingStaticRoutes = new HashMap<>();
+
+    @Nullable
+    private ChannelHandlerContext ctx;
+
+    record PendingResponse(CompletableFuture<ApiMessage> future, String route, NodeIdMapping nodeIdMapping) {}
+
+    enum ResponseOutcome {
+        CONSUMED,
+        STATIC_TRANSLATED,
+        UNHANDLED
+    }
+
+    RouteDispatcher(Map<String, RouteDescriptor> routes,
+                    NodeIdMapping nodeIdMapping,
+                    String routePrefix,
+                    CorrelationIdAllocator correlationIdAllocator,
+                    Map<Integer, HostPort> routerNodeAddresses,
+                    String virtualClusterName) {
+        this.routes = routes;
+        this.nodeIdMapping = nodeIdMapping;
+        this.routePrefix = routePrefix;
+        this.correlationIdAllocator = correlationIdAllocator;
+        this.routerNodeAddresses = routerNodeAddresses;
+        this.virtualClusterName = virtualClusterName;
+    }
+
+    void setContext(ChannelHandlerContext ctx) {
+        this.ctx = ctx;
+    }
+
+    @Override
+    public Map<String, RouteDescriptor> routes() {
+        return routes;
+    }
+
+    @Override
+    public NodeIdMapping nodeIdMapping() {
+        return nodeIdMapping;
+    }
+
+    public CorrelationIdAllocator correlationIdAllocator() {
+        return correlationIdAllocator;
+    }
+
+    public Map<Integer, HostPort> routerNodeAddresses() {
+        return routerNodeAddresses;
+    }
+
+    public Optional<HostPort> resolveRouterNodeAddress(int virtualNodeId) {
+        return Optional.ofNullable(routerNodeAddresses.get(virtualNodeId));
+    }
+
+    String qualifyRoute(String route) {
+        return routePrefix + route;
+    }
+
+    void trackStaticRoute(int correlationId, String route) {
+        pendingStaticRoutes.put(correlationId, route);
+    }
+
+    // --- Dispatch ---
+
+    @Override
+    public CompletionStage<ApiMessage> sendToAnyNode(String route,
+                                                     RequestHeaderData header,
+                                                     ApiMessage request,
+                                                     String sessionId,
+                                                     int clientCorrelationId) {
+        return executeOnEventLoop(() -> doSendToAnyNode(route, header, request, sessionId, clientCorrelationId));
+    }
+
+    @Override
+    public CompletionStage<ApiMessage> sendToSpecificNode(int targetNodeId,
+                                                          String route,
+                                                          RequestHeaderData header,
+                                                          ApiMessage request,
+                                                          String sessionId,
+                                                          int clientCorrelationId) {
+        return executeOnEventLoop(() -> doSendToSpecificNode(targetNodeId, route, header, request, sessionId, clientCorrelationId));
+    }
+
+    private CompletableFuture<ApiMessage> doSendToAnyNode(String route, RequestHeaderData header, ApiMessage request, String sessionId,
+                                                          int clientCorrelationId) {
+        RouteDescriptor rd = routes.get(route);
+        if (rd == null) {
+            withSendContext(LOGGER.atWarn(), sessionId, route, clientCorrelationId)
+                    .log("Router attempted to send to unknown route");
+            return CompletableFuture.failedFuture(new IllegalArgumentException("Unknown route: " + route));
+        }
+        String qualifiedRoute = qualifyRoute(route);
+        short requestApiVersion = header.requestApiVersion();
+        int routingCorrelationId = correlationIdAllocator.allocateId();
+        var frame = new DecodedRequestFrame<>(requestApiVersion, routingCorrelationId, true, header, request);
+        frame.setRouteName(qualifiedRoute);
+
+        if (!frame.hasResponse()) {
+            fireChannelRead(frame);
+            withSendContext(LOGGER.atTrace(), sessionId, route, clientCorrelationId)
+                    .addKeyValue("routingCorrelationId", routingCorrelationId)
+                    .log("Fire-and-forget request sent to route (no response expected)");
+            return CompletableFuture.completedFuture(null);
+        }
+
+        CompletableFuture<ApiMessage> future = new CompletableFuture<>();
+        pendingResponses.put(routingCorrelationId, new PendingResponse(future, route, nodeIdMapping));
+        fireChannelRead(frame);
+
+        withSendContext(LOGGER.atTrace(), sessionId, route, clientCorrelationId)
+                .addKeyValue("routingCorrelationId", routingCorrelationId)
+                .addKeyValue("apiVersion", requestApiVersion)
+                .log("Request sent to route");
+        return future;
+    }
+
+    private CompletableFuture<ApiMessage> doSendToSpecificNode(int targetNodeId,
+                                                               String route,
+                                                               RequestHeaderData header,
+                                                               ApiMessage request,
+                                                               String sessionId,
+                                                               int clientCorrelationId) {
+        RouteDescriptor rd = routes.get(route);
+        if (rd == null) {
+            withNodeContext(LOGGER.atWarn(), sessionId, route, targetNodeId)
+                    .log("Target node resolved to unknown route");
+            return CompletableFuture.failedFuture(
+                    new IllegalStateException("Node " + targetNodeId + " resolved to unknown route: " + route));
+        }
+
+        String qualifiedRoute = qualifyRoute(route);
+        short requestApiVersion = header.requestApiVersion();
+        int routingCorrelationId = correlationIdAllocator.allocateId();
+        var frame = new DecodedRequestFrame<>(requestApiVersion, routingCorrelationId, true, header, request);
+        frame.setRouteName(qualifiedRoute);
+        if (rd.targetsCluster()) {
+            frame.setTargetVirtualNodeId(targetNodeId);
+        }
+
+        if (!frame.hasResponse()) {
+            fireChannelRead(frame);
+            withSendContext(LOGGER.atTrace(), sessionId, route, clientCorrelationId)
+                    .addKeyValue("targetNodeId", targetNodeId)
+                    .addKeyValue("routingCorrelationId", routingCorrelationId)
+                    .log("Fire-and-forget request sent to specific node (no response expected)");
+            return CompletableFuture.completedFuture(null);
+        }
+
+        CompletableFuture<ApiMessage> future = new CompletableFuture<>();
+        pendingResponses.put(routingCorrelationId, new PendingResponse(future, route, nodeIdMapping));
+        fireChannelRead(frame);
+
+        withSendContext(LOGGER.atTrace(), sessionId, route, clientCorrelationId)
+                .addKeyValue("targetNodeId", targetNodeId)
+                .addKeyValue("routingCorrelationId", routingCorrelationId)
+                .log("Request sent to specific node");
+        return future;
+    }
+
+    // --- Response correlation ---
+
+    ResponseOutcome handleResponse(DecodedResponseFrame<?> frame, String sessionId) {
+        int correlationId = frame.correlationId();
+        PendingResponse pending = pendingResponses.remove(correlationId);
+        if (pending != null) {
+            NodeIdResponseTranslator.translate(frame.body(), frame.apiVersion(),
+                    pending.nodeIdMapping(), pending.route());
+            cacheNodeAddressesIfMetadata(frame.body(), sessionId);
+            pending.future().complete(frame.body());
+            frame.release();
+            LOGGER.atTrace()
+                    .addKeyValue("virtualCluster", virtualClusterName)
+                    .addKeyValue("sessionId", sessionId)
+                    .addKeyValue("routingCorrelationId", correlationId)
+                    .log("Routed response matched to pending request");
+            return ResponseOutcome.CONSUMED;
+        }
+
+        String staticRoute = pendingStaticRoutes.remove(correlationId);
+        if (staticRoute != null) {
+            NodeIdResponseTranslator.translate(frame.body(), frame.apiVersion(),
+                    nodeIdMapping, staticRoute);
+            cacheNodeAddressesIfMetadata(frame.body(), sessionId);
+            return ResponseOutcome.STATIC_TRANSLATED;
+        }
+
+        return ResponseOutcome.UNHANDLED;
+    }
+
+    // --- Lifecycle ---
+
+    void failAllPending(String sessionId) {
+        int abandoned = pendingResponses.size();
+        if (abandoned > 0) {
+            var cause = new IllegalStateException("Connection closed with " + abandoned + " pending router response(s)");
+            for (var entry : pendingResponses.values()) {
+                entry.future().completeExceptionally(cause);
+            }
+            pendingResponses.clear();
+            LOGGER.atWarn()
+                    .addKeyValue("virtualCluster", virtualClusterName)
+                    .addKeyValue("sessionId", sessionId)
+                    .addKeyValue("abandonedResponses", abandoned)
+                    .log("Connection closed with pending router responses");
+        }
+    }
+
+    // --- Internal helpers ---
+
+    // FutureReturnValueIgnored: a synchronous throw from work.get() is caught and
+    // propagated to `bridge`, so the submitted task cannot complete exceptionally and the
+    // whenComplete callback completes `bridge` in both branches.
+    @SuppressWarnings("FutureReturnValueIgnored")
+    private <T> CompletionStage<T> executeOnEventLoop(Supplier<CompletableFuture<T>> work) {
+        var executor = Objects.requireNonNull(ctx, "sendRequest called before handlerAdded").executor();
+        if (executor.inEventLoop()) {
+            return work.get();
+        }
+        CompletableFuture<T> bridge = new CompletableFuture<>();
+        executor.execute(() -> {
+            try {
+                work.get().whenComplete((r, e) -> {
+                    if (e != null) {
+                        bridge.completeExceptionally(e);
+                    }
+                    else {
+                        bridge.complete(r);
+                    }
+                });
+            }
+            catch (Exception e) {
+                bridge.completeExceptionally(e);
+            }
+        });
+        return bridge;
+    }
+
+    private void fireChannelRead(Object msg) {
+        Objects.requireNonNull(ctx, "fireChannelRead called before handlerAdded").fireChannelRead(msg);
+    }
+
+    private void cacheNodeAddressesIfMetadata(Object body, String sessionId) {
+        if (body instanceof MetadataResponseData md) {
+            for (var broker : md.brokers()) {
+                routerNodeAddresses.put(broker.nodeId(), new HostPort(broker.host(), broker.port()));
+            }
+            if (!md.brokers().isEmpty()) {
+                LOGGER.atDebug()
+                        .addKeyValue("virtualCluster", virtualClusterName)
+                        .addKeyValue("sessionId", sessionId)
+                        .addKeyValue("brokerCount", md.brokers().size())
+                        .log("Cached upstream node addresses from internal METADATA response");
+            }
+        }
+    }
+
+    private LoggingEventBuilder withSendContext(LoggingEventBuilder event, String sessionId, String route, int clientCorrelationId) {
+        return event.addKeyValue("virtualCluster", virtualClusterName)
+                .addKeyValue("sessionId", sessionId)
+                .addKeyValue("route", route)
+                .addKeyValue("clientCorrelationId", clientCorrelationId);
+    }
+
+    private LoggingEventBuilder withNodeContext(LoggingEventBuilder event, String sessionId, String route, int targetNodeId) {
+        return event.addKeyValue("virtualCluster", virtualClusterName)
+                .addKeyValue("sessionId", sessionId)
+                .addKeyValue("targetNodeId", targetNodeId)
+                .addKeyValue("route", route);
+    }
+}
